Added the star wars logo, and split Persons.js into multiple components for modularity
