--- src/components/persons/person-details/PersonDetails.js
@@ -1,7 +1,0 @@
-import React, { Component } from 'react';
-
-export default class PersonDetails extends Component {
-  render() {
-    return <div>Person Details, yo</div>;
-  }
-}
