+++ src/constants/index.js
@@ -1,0 +1,1 @@
+export const base_URL = `https://swapi.co/api`;
