+++ src/styles/Persons/persons.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-wrap: wrap;
   justify-content: space-around;
-  margin: 5em 0 2em 0;
+  margin: 3em 0 2em 0;
   position: absolute;
   z-index: -1;
 }
@@ -10,4 +10,11 @@
 .card-hover:hover {
   /* background: #ececec !important; */
   box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23) !important;
+}
+
+.logo {
+  fill: #ed4b6f;
+  max-width: 450px;
+  max-height: 75px;
+  margin: 0 auto;
 }
